+++ bank-feeds/src/codat_bankfeeds/models/shared/accountinfo.py
@@ -14,8 +14,7 @@ import pydantic
 from pydantic import model_serializer
 from pydantic.functional_serializers import PlainSerializer
 from pydantic.functional_validators import BeforeValidator
-from typing import TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class AccountInfoTypedDict(TypedDict):