+++ bank-feeds/src/codat_bankfeeds/models/shared/configuration.py
@@ -5,8 +5,8 @@ from .configurationschedule import ConfigurationSchedule, ConfigurationScheduleT
 from .syncconfiguration import SyncConfiguration, SyncConfigurationTypedDict
 from codat_bankfeeds.types import BaseModel
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class ConfigurationTypedDict(TypedDict):