+++ bank-feeds/src/codat_bankfeeds/models/shared/errorvalidation.py
@@ -10,8 +10,8 @@ from codat_bankfeeds.types import (
     UNSET_SENTINEL,
 )
 from pydantic import model_serializer
-from typing import List, TypedDict
-from typing_extensions import NotRequired
+from typing import List
+from typing_extensions import NotRequired, TypedDict
 
 
 class ErrorValidationTypedDict(TypedDict):